requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorzero==2.0
gpiozero==1.6.2
idna==2.10
logger==1.4
mailjet-rest==1.3.4
PyJWT==1.7.1
pynput==1.7.3
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Quartz==7.3
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sortedcontainers==2.4.0
twilio==6.61.0
urllib3==1.26.6
